Separate public and admin recruit controllers
The applicant-facing active recruitment endpoint should own the neutral RecruitController name, while existing /admin/recruits behavior remains explicitly marked as admin-only. This keeps URL contracts unchanged and postpones any broader admin API migration.

Constraint: Preserve existing /admin/recruits routes and the new /recruitments/active route

Rejected: Keep RecruitmentController | creates a naming split where the public recruit resource is not represented by RecruitController

Confidence: high

Scope-risk: narrow

Directive: Keep future applicant-facing recruitment APIs under RecruitController unless they are explicitly backoffice-only

Tested: ./gradlew test --tests 'org.ject.support.domain.recruit.controller.RecruitControllerTest' --tests 'org.ject.support.domain.recruit.service.RecruitServiceTest' --tests 'org.ject.support.domain.recruit.repository.RecruitRepositoryTest' -x jacocoTestCoverageVerification --rerun-tasks

Not-tested: AdminRecruitControllerTest execution; local Redis Testcontainers remains unavailable, but compileTestJava covered the renamed test and controller

Related: #534
