--- src/main/java/org/ject/support/domain/recruit/controller/RecruitmentController.java
@@ -1,22 +1,0 @@
-package org.ject.support.domain.recruit.controller;
-
-import lombok.RequiredArgsConstructor;
-import org.ject.support.domain.recruit.dto.ActiveRecruitmentResponses;
-import org.ject.support.domain.recruit.service.RecruitUsecase;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-
-@RestController
-@RequestMapping("/recruitments")
-@RequiredArgsConstructor
-public class RecruitmentController implements RecruitmentApiSpec {
-
-    private final RecruitUsecase recruitUsecase;
-
-    @Override
-    @GetMapping("/active")
-    public ActiveRecruitmentResponses findActiveRecruitments() {
-        return recruitUsecase.findActiveRecruitments();
-    }
-}
